Create README.md
Sun-Tracking-solar-panel

The demand for energy is increasing day by day. With the exhaustive use of non-renewable energy, we have finally landed on a platform of using renewable sources. What if we can optimize the usage of this energy too? Here is an innovative idea of using solar energy to it’s best possible extent. The basic idea of this project is to make solar panel rotate in all directions facing sunlight. It tracks the maximum intensity position and rests in that position. This maximizes the power per unit area and also increases the efficiency of the solar product systems. The demonstration involves AtMega 16 microcontroller enabled with a stepper motor. Based on the intensity of the light falling on the sensors, the dirrection of Sun is found. With the help of stepper motor, the panel is made to rotate in the direction of Sun thus enabling more intense dunlight to fall on the panel at any time of the day. This increases the power output.